Rename Purge and Res Tests
